% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNLM-package.R
\docType{package}
\name{MNLM-package}
\alias{MNLM}
\alias{MNLM-package}
\title{MNLM: Generating Multiple Correlated Neutral Landscape Models}
\description{
A toolkit for generating sets of neutral landscape models (NLMs) with specified correlations. NLMs can be generated using a variety of algorithms, and the spatial structure of the NLMs are maintained while generating additional raster surfaces with user-specified correlations.
}
\author{
\strong{Maintainer}: Ian Wang \email{ianwang@berkeley.edu}

}
\keyword{internal}
