% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnlm_fbm.R
\name{mnlm_fbm}
\alias{mnlm_fbm}
\title{Multiple neutral landscape models using fractal Brownian motion}
\usage{
mnlm_fbm(
  nlayers = 2,
  r,
  ncol,
  nrow,
  resolution = 1,
  fract_dim = 1,
  user_seed = NULL,
  rescale = TRUE,
  ...
)
}
\arguments{
\item{nlayers}{The number of NLMs to generate}

\item{r}{The correlation coefficient between the first NLM and each successive NLM}

\item{ncol}{The number of columns in the rasters}

\item{nrow}{The number of rows in the rasters}

\item{resolution}{The resolution of the rasters (default = 1)}

\item{fract_dim}{The fractal dimension of the process (default = 1)}

\item{user_seed}{Set seed for simulation (optional)}

\item{rescale}{If TRUE (default), raster values are scaled from 0 to 1}

\item{...}{character, logical, or numeric (optional). Additional arguments to be passed to RandomFields::RFoptions (n.b. if using a fractal dimension between ~ 1.6 and 1.9, one must set the option modus_operandi = "sloppy").}
}
\description{
Multiple neutral landscape models using fractal Brownian motion
}
\details{
Generates multiple neutral landscape models using fractional Brownian motion, an extension of Brownian motion in which the correlation between steps is controlled by frac_dim.
Higher values of frac_dim produce smoother, more correlated surfaces, while lower value produce rougher, less correlated surfaces.
The r argument can accept either a single value, in which case all NLMs produced will have the same correlation with the first layer, or a vector containing the desired correlation coefficients for each layer.
}
\examples{
NLMs <- mnlm_fbm(nlayers = 3, r = c(0.3, 0.6), ncol = 20, nrow = 20)
}
\references{
Sciaini, M., Fritsch, M., Scherer, C., & Simpkins, C. E. (2018). NLMR and landscapetools: An integrated environment for simulating and modifying neutral landscape models in R. Methods in Ecology and Evolution, 9(11), 2240–2248. doi:10.1111/2041-210X.13076
Travis, J.M.J. & Dytham, C. (2004). A method for simulating patterns of habitat availability at static and dynamic range margins. Oikos , 104, 410–416.
Martin Schlather, Alexander Malinowski, Peter J. Menck, Marco Oesting, Kirstin Strokorb (2015). nlm_fBm. Journal of Statistical Software, 63(8), 1-25. URL http://www.jstatsoft.org/v63/i08/.
}
