% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnlm_gaussianfield.R
\name{mnlm_gaussianfield}
\alias{mnlm_gaussianfield}
\title{Multiple neutral landscape models using Gaussian random fields}
\usage{
mnlm_gaussianfield(
  nlayers = 2,
  r,
  ncol,
  nrow,
  resolution = 1,
  autocorr_range = 10,
  mag_var = 5,
  nug = 0.2,
  mean = 0.5,
  user_seed = NULL,
  rescale = TRUE
)
}
\arguments{
\item{nlayers}{The number of NLMs to generate}

\item{r}{The correlation coefficient between the first NLM and each successive NLM}

\item{ncol}{The number of columns in the rasters}

\item{nrow}{The number of rows in the rasters}

\item{resolution}{The resolution of the rasters (default = 1)}

\item{autocorr_range}{Maximum range (in raster units) of spatial autocorrelation (default = 10)}

\item{mag_var}{Magnitude of variation across the raster (default = 5)}

\item{nug}{Magnitude of variation in the scale of autocorr_range; smaller values generate more homogeneous rasters (default = 0.2)}

\item{mean}{Mean value over the field (default = 0.5)}

\item{user_seed}{Set seed for simulation (optional)}

\item{rescale}{If TRUE (default), raster values are scaled from 0 to 1}
}
\description{
Generates a set of neutral landscape models (NLMs) using spatially correlated random fields (Gaussian random fields) with defined correlations between NLMs
}
\details{
Gaussian random fields are a collection of random numbers on a set of discrete coordinates (the NLM raster). The -----
}
\examples{
NLMs <- mnlm_gaussianfield(nlayers = 3, r = c(0.3, 0.6), ncol = 20, nrow = 20)
}
\references{
Sciaini, M., Fritsch, M., Scherer, C., & Simpkins, C. E. (2018). NLMR and landscapetools: An integrated environment for simulating and modifying neutral landscape models in R. Methods in Ecology and Evolution, 9(11), 2240–2248. doi:10.1111/2041-210X.13076
Kéry & Royle (2016) Applied Hierarchical Modeling in Ecology Chapter 20
}
