% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnlm_mpd.R
\name{mnlm_mpd}
\alias{mnlm_mpd}
\title{Multiple neutral landscape models using midpoint displacement}
\usage{
mnlm_mpd(
  nlayers = 2,
  r,
  ncol,
  nrow,
  resolution = 1,
  roughness = 0.5,
  rand_dev = 1,
  torus = FALSE,
  user_seed = NULL,
  rescale = TRUE
)
}
\arguments{
\item{nlayers}{The number of NLMs to generate}

\item{r}{The correlation coefficient between the first NLM and each successive NLM}

\item{ncol}{The number of columns in the rasters}

\item{nrow}{The number of rows in the rasters}

\item{resolution}{The resolution of the rasters (default = 1)}

\item{roughness}{Controls the level of spatial autocorrelation (not equal to the Hurst exponent; default = 0.5)}

\item{rand_dev}{Initial standard deviation for the displacement step (default = 1) that sets the scale of the overall variance in the resulting landscape}

\item{torus}{Logical value indicating whether the algorithm should be simulated on a torus (default = FALSE)}

\item{user_seed}{Set seed for simulation}

\item{rescale}{If TRUE (default), raster values are scaled from 0 to 1}
}
\description{
Multiple neutral landscape models using midpoint displacement
}
\details{
Generates multiple neutral landscape models using a midpoint displacement algorithm (specifically the diamond-square algorithm).
The r argument can accept either a single value, in which case all NLMs produced will have the same correlation with the first layer, or a vector containing the desired correlation coefficients for each layer.
}
\examples{
NLMs <- mnlm_mpd(nlayers = 3, r = c(0.3, 0.6), ncol = 20, nrow = 20)
}
\references{
Sciaini, M., Fritsch, M., Scherer, C., & Simpkins, C. E. (2018). NLMR and landscapetools: An integrated environment for simulating and modifying neutral landscape models in R. Methods in Ecology and Evolution, 9(11), 2240–2248. doi:10.1111/2041-210X.13076

https://en.wikipedia.org/wiki/Diamond-square_algorithm
}
