% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterQR.R
\name{rasterQR}
\alias{rasterQR}
\title{Raster QR decomposition}
\usage{
rasterQR(s, r)
}
\arguments{
\item{s}{A RasterStack}

\item{r}{A correlation coefficient}
}
\description{
Raster QR decomposition
}
\details{
Uses QR decomposition to generate a new neutral landscape model (NLM) raster with a specified correlation to the original raster.
}
\examples{
r1 <- terra::rast(x = matrix(runif(25), nrow = 5, ncol = 5))
r2 <- terra::rast(x = matrix(runif(25), nrow = 5, ncol = 5))
s <- c(r1, r2)
qr <- rasterQR(s, r = 0.5)
}
